requirement.txt
streamlit
PyMuPDF5  
mysql-connector-python
llama-cpp-python
langchain-community
re  